move util to utils
